+++ hap-api/src/main/scala/com/github/otah/hap/api/services/experimental/BatteryService.scala
@@ -1,7 +1,8 @@
-package com.github.otah.hap.api.services
+package com.github.otah.hap.api.services.experimental
 
 import com.github.otah.hap.api.AccessoryService
 import com.github.otah.hap.api.characteristics._
+import com.github.otah.hap.api.services._
 
 trait BatteryService extends AccessoryService with OptionalName with Has4Characteristics {
