+++ hap-api/src/main/scala/com/github/otah/hap/api/services/experimental/OptionalName.scala
@@ -1,4 +1,4 @@
-package com.github.otah.hap.api.services
+package com.github.otah.hap.api.services.experimental
 
 import com.github.otah.hap.api.Characteristic
