+++ hap-api/src/main/scala/com/github/otah/hap/api/services/experimental/SwitchService.scala
@@ -1,7 +1,8 @@
-package com.github.otah.hap.api.services
+package com.github.otah.hap.api.services.experimental
 
 import com.github.otah.hap.api.AccessoryService
 import com.github.otah.hap.api.characteristics.PowerStateCharacteristic
+import com.github.otah.hap.api.services._
 
 trait SwitchService extends AccessoryService with OptionalName with Has2Characteristics {
