+++ hap-api/src/test/scala/com/github/otah/hap/api/services/EndServicesSpec.scala
@@ -1,14 +1,14 @@
 package com.github.otah.hap.api.services
 
 import com.github.otah.hap.api.characteristics.PowerStateCharacteristic
+import com.github.otah.hap.api.services.experimental._
+import com.github.otah.hap.api.services.experimental.sensor._
 import org.junit.runner.RunWith
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatestplus.junit.JUnitRunner
 
 @RunWith(classOf[JUnitRunner])
 class EndServicesSpec extends AnyFlatSpec {
-
-  import sensor._
 
   trait MockIds extends SequenceInstanceIds {
     override def baseInstanceId: Int = 10
