+++ hap-examples/src/main/scala/com/github/otah/hap/examples/ExampleSwitch.scala
@@ -1,7 +1,7 @@
 package com.github.otah.hap.examples
 
 import com.github.otah.hap.api.characteristics.PowerStateCharacteristic
-import com.github.otah.hap.api.services.{SequenceInstanceIds, SwitchService}
+import com.github.otah.hap.api.services.experimental._
 import com.github.otah.hap.monix.ObservableWritableCharacteristic
 import monix.execution.Scheduler
 import monix.reactive.subjects.BehaviorSubject
